Added getNode function to linkPath
With this function it is possible to get the associated node from
a linkPath to get configuration parameters. This is to make the
configuratioin easier and less verbose.

Before this change a user might change the id of a link, but
not the request data which would leed to a visual update but the
same data being requested.